refactor: introduce dataClasses for ClusteringConfig and SpeciesConfig, update type annotations
